requirements.txt
GitPython==3.1.31
hydra-core==1.3.2
mlflow==2.14.3
numpy==1.25.2
numpy==1.25.0
omegaconf==2.3.0
optuna==3.6.1
pandas==2.0.1
pandas==2.0.3
pydantic==1.10.8
pytest==8.3.1
scikit_learn==1.2.2
setuptools==65.6.3
torch==2.0.1
tqdm==4.65.0
rtdl_num_embeddings==0.0.10